refactor(results): split types into common and results-specific files (#137)
